Extracted benchmarks into multiple files
